backend/requirements.txt
Django
djangorestframework
pillow
tensorflow
